Feat : ✨ 오늘의 조회수 API, 조회수 00시에 초기화

+++ server/src/main/java/com/cheerup/moomul/domain/member/entity/ProfileResponseDto.java
@@ -7,6 +7,6 @@ public record ProfileResponseDto(
 	Boolean isMine,
 	Long toMe,
 	Long fromMe,
-	Long today
+	int today
 ) {
 }
